pass db connection to loader, insert products

+++ models/product.go
@@ -4,7 +4,7 @@ type Product struct {
 	ID        string `gorm:"primaryKey"`
 	Name      string
 	Amount    int
-	Price     int
+	Price     float64
 	Materials []Material `gorm:"many2many:product_materials;"`
 }
